Introduced specialized `AttributeDictionary`
Ideally, models will optionally accept a dictionary of attributes so that scalar properties can be quickly and easily assigned without relying on reflection, in order to improve performance. The `AttributeDictionary` class will not only provide a friendlier signature than e.g. `IDictionary<string, string>`, but will also provide a friendlier `GetValue()` method for getting a value if there's a match, and otherwise returning `null`.

This will aid in supporting Feature #99.