Added `AttributeDictionary` constructors to each of the view models
The models in `OnTopic.ViewModels` are intended to be the standard reference models used for out-of-the-box implementations of OnTopic, and should be optimized for readability and performance, while also demonstrating best practices. To this end, I've added support for the new `AttributeDictionary` constructor to each of these, thus allowing them to bypass reflection for standard scalar properties.

Equally importantly, while not all implementors will choose to create `AttributeDictionary` constructors for their own view models, we want to support those who do, and do so by basing their implementation on the `OnTopic.ViewModels` library. By ensuring there's an `AttributeDictionary` constructor on every model—even models that aren't used very often and might not otherwise benefit from this functionality—we make it possible for any model in `OnTopic.ViewModels` to be subclassed and to have an `AttributeDictionary` constructor defined on that derived type without needing to implement properties of all base types.

This should improve both performance of the `OnTopic.ViewModels` when used directly. while offering flexibility for implementors to fully adopt `AttributeDictionary` constructors on their own derived view models, should they choose to do so.

This completes the basic implementation of the `AttributeDictionary` support for the `TopicMappingService` as required by Feature #99. That said, we still need to extend our unit tests to cover `AttributeDictionary` and the new `MapAsync()` functionality.