Updated `PropertyConfiguration` to accept `MemberAccessor`
Instead of passing a `PropertyInfo` to `PropertyConfiguration`, we now pass a `MemberAccessor`. This provides access to additional information, and will prevent dependents on `PropertyConfiguration` from needing to lookup a `MemberAccessor` via the `TypeAccessorCache`.

As part of this, I updated both `TopicMappingService` and `ReverseTopicMappingService` to pass the `MemberAccessor` to the constructor of `PropertyConfiguration`.