Updated `SetCollectionValueAsync()` to use `MemberAccessor()`
Previously, the `SetCollectionValueAsync()` was operating directly off of the `PropertyInfo`, instead of using the `MemberAccessor`, since it only had access to the `PropertyConfiguration` (and, thus, `PropertyInfo`). As a result, the calls to `SetValue()` and `GetValue()` were going directly through reflection, and not using the optimizations built into the new `MemberAccessor`.

With `MemberAccessor` now exposed as a property on `PropertyConfiguration` (49ffd613), we can easily remedy this. This should be a bit more performant, and at minimum helps ensure that all reflection access is going through `MemberAccessor`.

There remain similar gaps here in `ReverseTopicMappingService` and its dependency, `BindingModelValidator`; those will be handled in future commits.