Updated `BindingModelValidator` to use `MemberAccessor`
Previously, the `BindingModelValidator` was operating directly off of the `PropertyInfo`, instead of using the `MemberAccessor`, since most of the methods only had access to the `PropertyInfo` . As a result, the calls to `SetValue()` and `GetValue()` were going directly through reflection, and not using the optimizations built into the new `MemberAccessor`.

This behavior was a holdover from the legacy `TypeMemberInfoCollection` and `MemberDispatcher` architecture. While `BindingModelValidator` doesn't have any dependencies on those classes, it did continue to operate off of `PropertyInfo`.

Now, for consistency, I've updated `ValidateModel()` to simply accept a `TypeAccessor` instead of a `sourceType` and a `properties` list. Further, all of its private methods now accept a `MemberAccessor` instead of a `PropertyInfo`.

In addition to being more consistent with other libraries, this should also be a bit more performant, and at minimum helps ensure that all reflection access is going through `MemberAccessor`.

This also mitigates a dependency on `PropertyConfiguration` needed to support the new constructor signature, which accepts a `MemberAccessor` (49ffd613).