Updated `ItemConfiguration` to operate off of `List<Attribute>`
Previously, the `ItemConfiguration` operated off of a `ICustomAttributeProvider`, which gave it access to `GetCustomAttributes()`. This necessitated it calling `GetCustomAttributes()` at least once per property per request. Worse, due to inefficiencies in the code, it actually called it repeatedly.

In practice, however, these will never change per property—unlike other aspects of `ItemConfiguration` which are state dependent—and, thus, can be called from a cached source. With the new `CustomAttributes` property on `MemberAccessor` (4f587df4) and the `PropertyConfiguration` class now operating off of a `MemberAccessor` instead of a `PropertyInfo`, we can easily relay a `List<Attribute>` to `ItemConfiguration` instead of simply a `ParameterInfo` or `PropertyInfo` (which both implement `ICustomAttributeProvider`).

As part of this, I now set the `CustomAttributes` to a protected property of `PropertyConfiguration` for easy access to these values, so it doesn't need to be relayed internally, and can be accessed by derived classes (which will come in handy for a final optimization of `PropertyConfiguration`).