Update `Validate()` to use `CustomAttributes`
With the `CustomAttributes` now exposed to the underlying `ItemConfiguration` as a protected property (9705824a), the `PropertyConfiguration.Validate()` method can now access those instead of making a direct call to `GetCustomAttributes()`. This also updates the code to use the `MemberAccessor` version of `GetValue()` instead of going directly through reflection, which should have a slight performance benefit.

Technically, this could have been done before when the dependency on `MemberAccessor` was introduced to `PropertyConfiguration` (49ffd613). As we hadn't updated `ItemConfiguration` to operate off of `CustomAttributes` yet, however, I waited to migrate this so they could operate off of the same base property (9705824a).