Introduced unit tests for the new `AsAttributeDictionary()` method
The `AtAttributeDictionary()` method is a convenience method on `AttributeCollection` that allows all `AttributeRecord` objects in both the collection as well as, optionally, the collections of any `BaseTopic`  that the current topic derives from, to be converted into a lightweight `AttributeDictionary` (a7813565).

This commit introduces unit tests that verify that a) inheritance works as expected, and b) excluded attributes (such as `Title` and `LastModified`) are not set.