Introduced unit test for mapping constructors with an `AttributeDictionary`
As per Feature #99, we now support mapping models that have a constructor with a single `AttributeDictionary` parameter (72ca0324). In these cases, scalar properties in the `AttributeCollection` will be set via constructor, thus bypassing unnecessary calls to reflection. The new unit test confirms that this works by ensuring that such a constructor correctly sets a `Subtitle` property, without interfering with the standard handling for `Title` and `LastModified` (which are handled via compatible properties).