Established new `ItemMetadata` abstract class
In preparation for introducing a new `ParameterMetadata` class, I've moved common properties from `MemberAccessor` to a new abstract class, `ItemMetadata`. This includes `Type`, `IsNullable`, and `CustomAttributes`. With this in place, we'll be able to base `ItemConfiguration` off of `ItemMetadata` instead of (just) an `IEnumerable<Attribute>`, thus providing access to more metadata. This will also allow us to cache common metadata for parameters on the `TypeAccessor`, just as we do for members.

The `Type` treatment is a bit goofy because determining the type for property members requires some logic that's specific to whether it's a `MethodInfo` or a `PropertyInfo`. Since this makes most sense in `MemberAccessor`, the `Type` class is given an `init` method so that it can be set by derived classes. If this is not done, however, classes will not behave as expected. As an `abstract` class, this should be easy to ensure; if this were a `public` class, we'd take a different approach to avoid inadvertent errors. An `ArgumentNullException` will be thrown if this is not done.

As part of this, updated XML Docs to maintain the references to these base properties.