Introduced new `ParameterMetadata` class
With the abstract `ItemMetadata` in place (bec7123d), we can now create a new `ParameterMetadata` class, which shares a number of properties with the existing `MemberAccessor` class. This will, eventually, allow the `ItemConfiguration` to accept an `ItemMetadata` class for _either_ parameters _or_ properties, thus giving it access to more configuration data than simply `CustomAttributes`, as it has today. Ultimately, this will prevent us from needing to dynamically check e.g. `IsCompatible` and `IsList` by allowing those values to be added and cached as part of `ItemMetadata.