Introduced new `ConstructorParameters` property to `TypeAccessor`
With the new `ParameterMetadata` class established (37eeb3c8), we can now expose a collection of `ParameterMetadata` instances on `TypeClass` to represent the list of parameters on the primary (currently first) constructor. This way, this information is now cached, which will ultimately allow us to only call `GetPrimaryConstructor()` once per class, instead of every time an instance of that class is evaluated. In a subsequent commit, this will be incorporated into the `TopicMappingService`.