Updated `TopicMappingService` to use new `ParameterMetadata`
In `TopicMappingService`, when evaluating parameters, we now pull a list of `ParameterMetadata` from the `TypeAccessor`, instead of calling `GetPrimaryConstructor()` for each instance. This should provide a slight improvement to performance. Outside of that, this doesn't buy us too much yet—though it will be get us more functionality in a bit, once we incorporate it into `ItemConfiguration` as well. That'll come in a subsequent commit.

As part of this, this required updating the parameters and member calls to work off of `ParameterMetadata` instead of `ParameterInfo`.