Update `ItemConfiguration` to depend on new `ItemMetadata`
Previously, the `ItemConfiguration` accepted an `IEnumerable<Attribute>` (9705824a), which is the common data used for `PropertyInfo`, `MethodInfo`, and `ParameterInfo`. Now, with the establishing of the `ItemMetadata` abstract class (bec7123d), which is shared between `MemberAccessor` and the new `ParameterMetadata` (37eeb3c8), we can instead accept an `ItemMetadata`. This gives us access to the `CustomAttributes`, just as before, but also provides access to additional metadata.

This will allow better sharing (and caching) of state data in `TopicMappingService` by allowing properties to be assessed off of `ItemMetadata` instead of reevaluating reflection data for every call. That will come in subsequent commits.