Moved `IsList` and `IsConvertible` to `ItemMetadata` properties
With `ItemMetadata` incorporated into `ItemConfiguration`, any method which relies on `ItemConfiguration` or `PropertyConfiguration` now has access to metadata that's shared between parameters (via `ParameterMetadata`) or methods or properties (via `MemberAccessor`). Given that, we can move `IsList()` to `ItemMetadata`, and cache the values of both `IsList` and `IsConvertible` via properties on `ItemMetadata`.

This is useful as it means we can evaluate these once upfront, instead of evaluating them each time we evaluate a property. These lookups were pretty cheap—they primarily involve a `Contains()` check on a `List<Type>`—but needing to do this at least once for every property during `MapAsync()` adds some overhead. Since we already have a cache of `ItemMetadata` via `TypeAccessor` (and the `TypeAccessorCache`), it's easy to include that in the cached metadata.