Established cached version of `TypeAccessor` for `Topic`
When evaluating members of the source `Topic` class in `TopicMappingService`, we are almost always operating off of an actual `Topic` class, since subclasses of `Topic` are rare, and mostly used for configuration data. Given that, it doesn't make sense to do a lookup of the `TypeAccessor` for the `Topic` class via the `TypeAccessorCache` every time we need to do this (e.g., as part of `GetValue()` and `TryGetCompatibleValue()`. This lookup was probably fast, but there's no need to repeat it when we have a logical default.