Reorder `GetValue()` based on how common each condition is
Previously, `GetValue()` was ordered according to popularity (with the exception of the compatibility check, which must come first). These were recently reordered to make sure that the fastest calls were placed upfront, thus preventing lookups when they weren't needed (565fa6fc). With the caching of `IsList` and `IsConvertible` (d0fb8a48), however, that is no longer a concern. As such, we can return to prioritizing based on how common each is, thus reducing the number of conditions each property (or parameter) must be considered for.