Established basic load test for `TopicMappingService`
We should establish more sophisticated load testing. In the meanwhile, this unit test helps provide basic load testing of mapping calls by configuring a number of variables, such as the number of runs, whether or not a comprehensive attribute set should be used, and whether nested topics should be included. This allows optimizations to the code base to be quickly evaluated. This test always passes; the goal is to measure the time it takes. It defaults to 0 runs so it doesn't slow down normal unit test operations.