Established basic test for evaluating thresholds on `TopicMappingService`
This unit test and its corresponding `LoadTestingViewModel` allows a variable number of attributes to be set on a `Topic` which correspond to upwards of 20 properties on the `LoadTestingViewModel`. The number of properties to be mapped and the number of times to map the `Topic` are both variables to be set. These allow us to help identify and optimize the threshold to be used in the `TopicMappingService` for when to utilize the `AttributeDictionary`.