Established initial threshold for `AttributeDictionary` mapping
The `AttributeDictionary` mapping introduces some overhead. If there aren't many attributes that will be mapped, this will actually slow down the mapping service, instead of speeding it up. If there are a lot of attributes that are mapped, however, it will dramatically improve the performance. With the help of the new load tests (3ecfa069, 7b529dc6), I determined that this threshold is pretty quick; if 2-3 attributes are mapped, we reach a breakeven point; if more than 5-6 attributes are mapped, we start seeing a significant reduction.

Unfortunately, predetermining if the attributes map to properties is itself expensive, so we're instead just establishing a heuristic. In a typical OnTopic implementation, we typically expect each topic to have 2-3 attributes that won't be mapped with `AttributeDictionary` (such as `Title`, `LastModified`, `LastModifiedBy`). So the assumption here is that if there are five or more attributes, then 2-3 will be mapped, thus reaching that breakeven point.

In addition, if those attributes don't have properties to map to, there isn't any benefit to passing them in via the `AttributeDictionary`. This is even harder to evaluate because many of those will be compatible properties that aren't sourced from attributes (such as `Id`, `Key`, `ContentType`, `WebPath`, &c.)—but that will depend on the model itself, and may not hold for e.g. nested topics or references, or `[MapAs()]` types. In fact, we expect nearly all models to have 5-7 "base" properties as a result of that, in order to maintain compatibility with `ITopicViewModel`. Nevertheless, if there are less than five compatible properties, this likely isn't going to be worth the overhead.

In the future, we may revise this further based on real-world telemetry. Further, we may want to e.g. cache the number of convertible properties on the `TypeAccessor`, and even attempt to estimate the number of non-mapped types based on e.g., whether the target model implements `ITopicViewModel`. That's a bit expensive of a check, but again can be cached as part of the `TypeAccessor.