Updated `Host` sites to disable `includeStaticFiles`
By passing the `includeStaticFiles` argument in `MapTopicErrors()` (b591bd7c), we are effectively disabling `includeStaticFiles` in the corresponding `ErrorController.Http()` action (83f4691f). While `includeStaticFiles` defaults to true, since it's the intuitive baseline, we expect most implementers will want to disable it, and thus have made it the standard in the boilerplate `Host` file. Adding it to the integration tests' `Host` will allow us to follow-up with an integration test of the Feature #101 functionality; see next commit.