Introduced integration test for new `includeStaticFiles` parameter
Confirm that by setting the `includeStaticFiles` argument in `MapTopicErrors()` (b591bd7c) for the integration tests host (74b7cc8b), we are effectively disabling `includeStaticFiles` in the corresponding `ErrorController.Http()` action (83f4691f).

As part of this, I needed to remove the (otherwise extraneous) check for the correct content type, since the custom error for static files will be of type `text/plain` instead of `text/html`.

This validates the functionality called for in Feature #101.