Add `ItemConfiguration` as property to `ItemMetadata`
The majority of the `ItemConfiguration` properties represent attributes (and their values) associated with individual parameters, properties, or methods, and whose values will not change at runtime. Given that, reidentifying these for every property incurs some overhead, and especially when we already have an `ItemMetadata` class for caching this information (via `TypeAccessor` and `TypeAccessorCache`).

The one exception to this is the `attributePrefix` constructor argument, which conditionally updates the `AttributeKey` property. That is passed to calls to e.g. the `TopicMappingService` and `ReverseTopicMappingService`. In subsequent updates, that will be removed from the `ItemConfiguration`, and managed internally within those classes.