Move `Validate()` method to `MemberAccessor`
In preparation for centralizing around `ItemConfiguration`, we need to move the `Validate()` method from `PropertyConfiguration` to `MemberAccessor`.

This is because it relies on access to the `MemberAccessor`'s `GetValue()` method. Currently, `PropertyConfiguration` has access to that via its `MemberAccessor` property. But once `ItemConfiguration` becomes a property off of `ItemMetadata`, it doesn't make sense to maintain the `ItemMetadata` or `MemberAccessor` fields off of `ItemConfiguration` or `PropertyConfiguration`.

Instead, callers will simply pass references to `ItemMetadata` (which will include `ItemConfiguration`) instead of `ItemConfiguration` (which previously included `ItemMetadata`). Basically, we're flipping the relationship between these.

Further, this can't be added to the `ItemMetadata` base class since it requires access to the `GetValue()` method on the derived `MemberAccessor` class. That's the same reason it was previously added to the `PropertyConfiguration`, not the `ItemConfiguration` class.