Migrate mapping services to use `ItemMetadata.Configuration`
The `TopicMappingService` and `ReverseTopicMappingService` (and, by extension, the `BindingModelValidator`) have been updated to pull the `ItemConfiguration` from `ItemMetadata` (or `PropertyMetadata`) (74cb8173) instead of constructing a new version directly. This takes advantage of the caching of `TypeAccessor` and `ParameterMetadata`, thus ensuring the configuration data only needs to be assessed once per item.

This loses access to the `attributePrefix` state data, and thus temporarily breaks the `[MapToParent]` functionality. In a future commit, we'll update the `TopicMappingService` and `ReverseTopicMappingService` to relay that internally, and reference it as needed.

This may also temporarily break some references that were relying on specific properties of `PropertyConfiguration`, namely `MemberAccessor`; that will be resolved in a subsequent commit, when we begin relaying `MemberAccessor` instead of `PropertyConfiguration`.