Remove `attributePrefix` from `ItemConfiguration`
With the `ItemConfiguration` moved to a `Configuration` property off of `ItemMetadata` (74cb8173, 2f482b12), it's no longer receiving runtime data on the contextual configuration needed to populate the `attributePrefix`. In subsequent commits, we'll be updating the `TopicMappingService` and `ReverseTopicMappingService` to pass these internally. In the meanwhile, it can safely be removed from the `ItemConfiguration` since it's not being actively used, and the parent `ItemMetadata` class doesn't have access to it regardless.

I also removed the `name` parameter since that can be pulled directly from `itemMetadata` parameter.