Relay `ItemMetadata`, `attributePrefix` instead of `ItemConfiguration`
With `ItemConfiguration` now a property on `ItemMetadata` (74cb8173), it doesn't make sense to pass `ItemConfiguration` from method-to-method, but instead pass `ItemMetadata` (or `MemberAccessor` if specifically operating off of properties and methods). This effectively flips the order of dependencies, with `ItemConfiguration` being a property of `ItemMetadata` instead of `ItemMetadata` being a property of `Configuration`.

In addition, I now pass the `attributePrefix` from method-to-method as well. This used to be part of the `ItemConfiguration`. Since it was the one (and only) runtime variable included in that, and since `ItemConfiguration` is now cached as part of the `TypeAccessorCache` hierarchy, the `attributePrefix` was removed from `ItemConfiguration`, thus breaking the `[MapToParent()]` functionality (84735fbb). By passing it from method-to-method, we take the first step toward fixing this.

As part of this, I also removed the `ItemMetadata` property from the `ItemConfiguration` class since it's no longer needed.