Introduced new `GetCompositeAttributeKey()` method
With the `attributePrefix` removed from the `ItemConfiguration`, the onus is on the caller to ensure that the `attributePrefix` is accounted for, if relevant. To simplify this process, I've added a simple convenience method for concatenating the full attribute key. This will be implemented in subsequent versions of `TopicMappingService` and `ReverseTopicMappingService`.

I've also updated all calls to `AttributeKey` to now call the new `GetCompositeAttributeKey()`. This effectively reinstates support for `attributePrefix`, which was broken when `ItemConfiguration` was moved to `ItemMetadata` (74cb8173, 84735fbb) and the mapping services were updated to rely on `ItemMetadata` instead of creating their own instances of `ItemConfiguration` (fe393021).

All unit tests related to `[MapToParent]` now pass again.