Removed legacy `PropertyConfiguration` class
With the `ItemConfiguration` being added as a property to `ItemMetadata` (74cb8173), and the mapping services being updated to rely upon `ItemMetadata` to get their configuration data (2f482b12, fe393021), there is no longer a need for the legacy `PropertyConfiguration` class, or to differentiate between item and property differentiation more broadly.

Indeed, the main benefit of this was to allow strongly typed access to the `MemberAccessor`, instead of general access to `ItemMetadata`. Since the order if flipped, with `ItemConfiguration` being retrieved _via_ `ItemMetadata` (or `MemberAccessor`), there is always strongly typed access to these objects independent of the configuration object.