Introduce `MaybeCompatible` property to `ItemMetadata`
The `MaybeCompatible` property attempts to determine whether a compatible or convertible member is available on a corresponding topic. This is useful as a quick heuristic for weeding out cases where compatibility is unexpected. Due to limitations of the implementation, it cannot guarantee compatibility—or incompatibility—since it doesn't know the source topic type. But, for our purposes, it should be good enough at indicating whether compatibility is likely.

The will provide the foundation for #Feature #102.