Introduce `SetItemCompatibility()` method in `TypeAccessor`
While the `MaybeCompatible` property is the main interface for determining compatibility when calling into `TypeAccessor` (6662e18e), the business logic for determining what may or may not be compatible is handled by `TypeAccessor`. The `SetItemCompatibility()` is a private helper method called by the constructor to set `MaybeCompatible` on both `MemberAccessor` and `ParameterMetadata` instances.

This addresses the core functionality of #102.