Exclude `record` members from `TypeAccessor`
The C# compiler automatically adds a number of members to `record` types, such as an `EqualityContract` and `ToString()`. These aren't relevant to either dispatch or mapping operations, and just add noise to the cache. These can be excluded. To ensure this works, I've converted `CustomTopicTopicViewModel` (53106550) to a `record` so its `MaybeCompatible` checks can validate whether there are any unexpected members.

While I was at it, I also consolidated all of the escape clauses in `IsValid()` into a single block for readability and succinctness.