Updated to operate off of `OnTopic.All` assembly
The `OnTopic.All` assembly contains one copy of all of the custom assemblies generated by the site. As such, it can be used as a single source of API metadata.

It's worth noting that since DocFx doesn't yet recognize C# 10's implicit or global `using` directives (dotnet/docfx#7737) we need to operate off of the assembly directly, instead of relying on either the project files (`*.csproj`) or source code (`*.cs`).