Establish Repositories filter
The `Repositories.yml` filter includes the `OnTopic.Repositories` namespaces under the `/Repositories` folder. As part of this, the `OnTopic.Repositories` namespaces are also removed from the `Base` filter (3a54aca7). This is a bit confusing in that it combines multiples assemblies into one section, but as these are all closely related to each other, it makes up for that in convenience and intuition of grouping.