Establish Mapping filter, section
The `Mapping.yml` filter includes the `OnTopic.Mapping` namespaces under a new `/Mapping` folder. As part of this, the `OnTopic.Mapping` namespaces are also removed from the `Base` filter (3a54aca7). This is useful as the mapping libraries are largely independent of the base libraries—so much so that they're a candidate for separation in OnTopic 6.x (#34)—and are also a major source of complexity. Having them in their own section makes intuitive sense from the implementer's perspective.