Ensured that View Model defaults aren't repeated in constructor
Previously, some view model defaults were set on both the property's auto-value, as well as the `AttributeDictionary` constructor. This ensures that the constructor instead falls back to the auto-value. While this isn't strictly required, it helps avoid scenarios where the default value is updated in one place, but accidentally overlooked in the other.