Add unique constraint for `TopicKey` and `ParentID`
This ensures that a topic with the same key as another topic cannot exist within the same parent scope. This is already enforced by the Topic Library itself, but not by the SQL schema. This helps prevent inadvertent conflicts from being introduced when manually updating the SQL database, which would result in an exception when loading the data via an `ITopicRepository` implementation. This resolves #106.