Renamed `AttributeDictionary` to `AttributeValueDictionary`
The name `AttributeDictionary` is more consistent with our overall naming conventions. However, it conflicts with the out-of-the-box `AttributeDictionary` class in the `Microsoft.AspNetCore.Mvc.ViewFeatures` namespace. As this is commonly introduced as a global namespace for model libraries, this introduces a potential conflict which requires fully-qualifying classes from either the corresponding Microsoft or the OnTopic namespace.

To mitigate that, I'm renaming `AttributeDictionary` to `AttributeValueDictionary`. This is similar to the out-of-the-box `RouteValueDictionary` which refers to route parameter keys and values, but uses the name `RouteValue`, not `RouteParameter` or even `Route`. That isn't an entirely satisfying precedent as it's pretty anomalous within the BCL, but it's also a popular enough class that the convention doesn't feel _entirely_ arbitrary, and especially by comparison to other options.