Revert: Renamed `AttributeDictionary` to `AttributeValueDictionary`
Renamed `AttributeValueDictionary` back to `AttributeDictionary`, thus reverting 38c54c39. It seems this decision was premature, and we can mitigate the naming conflicts without renaming the class, since the out-of-the-box `ViewFeatures` namespace is almost exclusively used in views, and the `AttributeDictionary` is almost exclusively used in view models. This effectively undoes #107.