IDE0079: Remove unnecessary IDE0060 suppressions
The IDE0060 is thrown when internal or private methods have parameters that are never used. Previously, this was thrown even for obsolete classes, where it's expected that the parameters won't be used.

Note: Removing these simply reintroduced the errors for `AttributeValue`, `AttributeValueCollection`, and `RelatedTopicCollection`. For those reasons, `IDE0060` was left in those files. We may want to reevaluate this later.