IDE0079: Remove unnecessary CA1024 suppressions
Previously, `CA1024` threw a warning anytime a public method started with `Get`, ended with a property name, and didn't accept a parameter, suggesting that a property be used instead. This occurred even if there was already a getter for the property, suggesting some additional need for the `Get{PropertyName}()` method—such as, in these cases, a comparatively expensive call. The rules for this appear to have been relaxed, allowing these cases.