IDE0079: Remove unnecessary CA1711 suppressions
The `CA1711` warning is thrown when e.g. a class name ends with `Collection` which isn't a `Collection`. That makes sense. In this case, however, the term "collection" is overloaded in xUnit, and thus the suppression. It isn't clear if an exception was made for cases where a class inherits from an interface containing the term `Collection`, or if a special exception was made for xUnit collections. Regardless, this is no longer needed, apparently.